backend/requirements.txt
# Core Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.18.0
seaborn>=0.11.0

# Web & Networking
websocket-client==1.6.4
websockets==12.0
requests>=2.31.0
aiohttp==3.9.1
fastapi==0.109.2
uvicorn==0.27.1

# Async & Threading
asyncio==3.4.3
aiofiles>=24.1.0

# AI/ML
torch==2.1.1
transformers>=4.46.0
langchain>=0.3.0
openai>=1.54.0
scikit-learn>=1.7.0

# Utilities
python-dotenv==1.0.0
pydantic==2.6.1
click==8.1.7
rich>=13.7.0
PyYAML>=6.0.1
toml>=0.10.2
jsonschema>=4.23.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Security
cryptography>=42.0.0
bcrypt>=4.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Streamlit & Web Interface
streamlit>=1.32.0
python-multipart==0.0.9

# System & Monitoring
psutil>=5.9.0
netifaces>=0.11.0
pynvml>=11.5.0

# Text Processing
textblob>=0.18.0
lxml>=5.3.0
beautifulsoup4>=4.12.0

# Computer Vision
opencv-python>=4.10.0
pillow>=10.0.0
imageio>=2.36.0
pytesseract>=0.3.10

# Network & Graph
networkx>=2.6.0
graphviz>=0.20.0

# Utilities
colorama==0.4.6
tqdm>=4.66.0
python-dateutil>=2.9.0
feedparser>=6.0.0
termcolor>=2.4.0
pyfiglet>=1.0.0
asciitree>=0.3.3

# HTTP & API
httpx>=0.28.0
httpcore>=1.0.0
h11==0.14.0

# Core Dependencies
anthropic>=0.52.0
annotated-types>=0.7.0
anyio==4.2.0
certifi>=2025.1.1
distro>=1.9.0
idna==3.6
jiter>=0.10.0
joblib>=1.5.0
markdown-it-py>=3.0.0
mdurl>=0.1.2
pydantic-core>=2.33.0
pygments>=2.19.0
sniffio==1.3.0
threadpoolctl>=3.6.0
typing-inspection>=0.4.0
typing-extensions==4.8.0
python-json-logger>=2.0.7
uuid7>=0.1.0
gitpython>=3.1.0
kaleido>=0.2.1

# Added from the code block
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0

# New dependencies from the code block
starlette==0.27.0
watchfiles==0.21.0
httptools==0.6.1

# Added from the code block
pygame>=2.5.0
